move test of _spark_script_args() to tests.test_runner
